[6477] - move teach firsrt migration service to within migration class
