+++ .rubocop.yml
@@ -107,3 +107,11 @@ Rails/RedundantActiveRecordAllMethod:
     - 'spec/lib/dqt/params/update_spec.rb'
     - 'spec/models/degree_spec.rb'
 
+RSpec/FilePath:
+  Exclude:
+    - 'spec/data_migrations/*'
+
+RSpec/SpecFilePathFormat:
+  Exclude:
+    - 'spec/data_migrations/*'
+
